% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulk.R
\name{pseudoBulk}
\alias{pseudoBulk}
\title{Aggregate single-cell data into pseudo-bulk counts}
\usage{
pseudoBulk(
  sc.obj,
  assay = "counts",
  sample,
  group = NULL,
  barcode = NULL,
  threshold = 2
)
}
\arguments{
\item{sc.obj}{A single-cell object in either Seurat or SingleCellExperiment format, containing count data and sample IDs.}

\item{assay}{The name of the assay to pull counts from, with "counts" as the default.}

\item{sample}{A string specifying the column in the metadata that contains sample information.}

\item{group}{Optional. A string specifying the column in the metadata that contains group information for differential expression.}

\item{barcode}{Optional. A vector of two elements where the first element is the name of the barcode column from the metadata,
and the second element is the specific barcode ID to filter cells by. This argument is used to subset cells based on barcode information.}

\item{threshold}{The minimum number of samples for each group to be included in the analysis.
A threshold less than 2 will cause the function to stop with an error.}
}
\value{
Returns a DGE object with normalised counts for a sample.
}
\description{
Aaggregates single-cell counts into pseudo-bulk counts per sample, 
and stores the pseudo-bulk counts in a DGEList object along with any metadata specified as groups for further differential expression analysis. 
If no barcode is specified, it aggregates all cells.
It enforces a minimum sample size per group as specified by the threshold.
}
\details{
pseudoBulk
}
